config.sh: Unify config.sh scripts across LTS and non-LTS builds.
Comparing the two config scripts, the only difference is that for LTS builds, a patch is omitted.

Assuming the patch does not apply, detect if the build targets LTS branch (through ENV we created in a previous PR, and gracefully proceed to build if true.

While we are at it, remove config-lts.sh and all mentions of it.

Signed-off-by: taalojarvi <sreedevan05@gmail.com>